+++ src/app/absence/planning-absence/planning-absence.component.html
@@ -1,0 +1,5 @@
+<div class="row justify-content-md-center">
+    <div class="col-md-auto" style="border: 1px solid #ced4da;background-color: #e3f2fd; border-radius:7px; ">
+        <h2>Planning des absences</h2>
+    </div>
+</div>
