+++ src/app/absence/planning-absence/planning-absence.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-planning-absence',
+  templateUrl: './planning-absence.component.html',
+  styleUrls: ['./planning-absence.component.scss']
+})
+export class PlanningAbsenceComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
