+++ src/app/app-routing.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { TechComponent } from './tech/tech.component';
 import { AuthComponent } from './auth/auth.component';
-import { ListerJourFerieComponent } from './jourFerie/lister-jour-ferie/lister-jour-ferie.component';
-import { CreationJourFerieComponent } from './jourFerie/creation-jour-ferie/creation-jour-ferie.component';
+import { ListerJourFermeComponent } from './jourFerme/lister-jour-ferme/lister-jour-ferme.component';
+import { CreationJourFermeComponent } from './jourFerme/creation-jour-ferme/creation-jour-ferme.component';
 import { StatutConnecteService } from './service/statut-connecte.service';
 import { StatutAdministrateurService } from './service/statut-administrateur.service';
 import { FourOhFourComponent } from './erreurNavigation/four-oh-four/four-oh-four.component';
@@ -11,11 +11,12 @@ import { AccesRefuseComponent } from './erreurNavigation/acces-refuse/acces-refu
 import { DemandeAbsenceComponent } from './absence/demande-absence/demande-absence.component';
 import { VisualisationAbsenceComponent } from './absence/visualisation-absence/visualisation-absence.component';
 import { ModificationAbsenceComponent } from './absence/modification-absence/modification-absence.component';
-import { ModificationJourFerieComponent } from './jourFerie/modification-jour-ferie/modification-jour-ferie.component';
+import { ModificationJourFermeComponent } from './jourFerme/modification-jour-ferme/modification-jour-ferme.component';
 import { VueSynthetiquesComponent } from './vue-synthetiques/vue-synthetiques.component';
 import { StatutManagerService } from './service/statut-manager.service';
 import { VueParDepartementParJourParCollaborteurComponent } from './vue-synthetiques/vue-par-departement-par-jour-par-collaborteur/vue-par-departement-par-jour-par-collaborteur.component';
 import { HistogrammeParDepartementParJourComponent } from './vue-synthetiques/histogramme-par-departement-par-jour/histogramme-par-departement-par-jour.component';
+import { PlanningAbsenceComponent } from './absence/planning-absence/planning-absence.component';
 
 const routes: Routes = [
   // canActivate: [StatutConnecteService] ==> Permet de verifier si l'utilisateur est connecte
@@ -24,13 +25,14 @@ const routes: Routes = [
   // canActivate: [StatutEmployeService] ==> Permet de verifier si l'utilisateur connecte est bien un employe
   { path: 'tech', component: TechComponent, canActivate: [StatutConnecteService] }, //  tech accessible uniquement si connecte
   { path: 'auth', component: AuthComponent },
-  { path: 'listerJourFerie', component: ListerJourFerieComponent, canActivate: [StatutConnecteService] },
-  { path: 'creationJourFerie', component: CreationJourFerieComponent, canActivate: [StatutConnecteService, StatutAdministrateurService] },
+  { path: 'listerJourFerme', component: ListerJourFermeComponent, canActivate: [StatutConnecteService] },
+  { path: 'creationJourFerme', component: CreationJourFermeComponent, canActivate: [StatutConnecteService, StatutAdministrateurService] },
   { path: 'accesRefuse', component: AccesRefuseComponent },
   { path: 'demandeAbsence', component: DemandeAbsenceComponent, canActivate: [StatutConnecteService] },
   { path: 'visualisationAbsence', component: VisualisationAbsenceComponent, canActivate: [StatutConnecteService] },
   { path: 'modificationAbsence/:id', component: ModificationAbsenceComponent, canActivate: [StatutConnecteService, StatutAdministrateurService] },
-  { path: 'modificationJourFerie/:id', component: ModificationJourFerieComponent, canActivate: [StatutConnecteService, StatutAdministrateurService] },
+  { path: 'modificationJourFerme/:id', component: ModificationJourFermeComponent, canActivate: [StatutConnecteService, StatutAdministrateurService] },
+  { path: 'planningAbsence', component: PlanningAbsenceComponent, canActivate: [StatutConnecteService] },
   { path: 'vueSynthetiques', component: VueSynthetiquesComponent, canActivate: [StatutConnecteService, StatutManagerService] },
   { path: 'vueParDepartementParJourParCollaborateur', component: VueParDepartementParJourParCollaborteurComponent, canActivate: [StatutConnecteService, StatutManagerService] },
   { path: 'histogrammeParDepartementParJour', component: HistogrammeParDepartementParJourComponent, canActivate: [StatutConnecteService, StatutManagerService] },
