+++ src/app/jourFerme/creation-jour-ferme/creation-jour-ferme.component.html
@@ -44,7 +44,7 @@
 					</button>
 
 					<!-- bouton annuler formulaire -->
-					<a routerLink="../listerJourFerie"><button class="btn btn-danger">
+					<a routerLink="../listerJourFerme"><button class="btn btn-danger">
 							<fa-icon [icon]="faTimes"></fa-icon> Annuler
 						</button>
 					</a>
