+++ src/app/jourFerme/lister-jour-ferme/lister-jour-ferme.component.scss
