+++ src/app/jourFerme/modification-jour-ferme/modification-jour-ferme.component.ts
@@ -1,47 +1,65 @@
 import { Component, OnInit } from '@angular/core';
-import { FormGroup, FormBuilder, Validators } from '@angular/forms';
-import { Router } from '@angular/router';
+import { Router, ActivatedRoute } from '@angular/router';
 import { JourFermeService } from 'src/app/service/jour-ferme.service';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { formatDate } from '@angular/common';
+import { JourFermeVisualisation } from 'src/app/models/jour-ferme-visualisation';
 import { faCheck, faTimes } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
-  selector: 'app-creation-jour-ferie',
-  templateUrl: './creation-jour-ferie.component.html',
-  styleUrls: ['./creation-jour-ferie.component.scss']
+  selector: 'app-modification-jour-ferme',
+  templateUrl: './modification-jour-ferme.component.html',
+  styleUrls: ['./modification-jour-ferme.component.scss']
 })
-export class CreationJourFerieComponent implements OnInit {
+export class ModificationJourFermeComponent implements OnInit {
 
   // Icones
   faCheck = faCheck;
   faTimes = faTimes;
 
   // Initialisations
-  formCreationJourFerme: FormGroup;
+  formModificationJourFerme: FormGroup;
   messageErreur = '';
   messageValidation = '';
+  id: number;
+  jourFerme: JourFermeVisualisation;
 
   // Constructeur
-  constructor(private router: Router, private formBuilder: FormBuilder, private jourFermeService: JourFermeService) { }
+  constructor(private router: Router,
+    private formBuilder: FormBuilder,
+    private jourFermeService: JourFermeService,
+    private routerLinkActive: ActivatedRoute) { }
 
   ngOnInit(): void {
-    this.initialiserFormulaire();
+    // Snapshot pour récupérer l'id passé via l'url
+    this.id = this.routerLinkActive.snapshot.params['id'];
+
+    // Subscription à l'observable
+    this.jourFermeService.getJourFermeParId(this.id).subscribe(
+      (jour) => {
+        this.jourFerme = jour;
+        this.initialiserFormulaire();
+      }, (error) => {
+        console.log('Erreur ' + error);
+      }
+    )
+
   }
 
   initialiserFormulaire() {
-    this.formCreationJourFerme = this.formBuilder.group({
-      dateJourFerme: ['', Validators.required],
-      typeJourFerme: ['', Validators.required],
-      commentaireJourFerme: ['']
+    this.formModificationJourFerme = this.formBuilder.group({
+      dateJourFerme: [this.jourFerme.date, Validators.required],
+      typeJourFerme: [this.jourFerme.type, Validators.required],
+      commentaireJourFerme: [this.jourFerme.commentaire]
     });
   }
 
   validerFormulaire() {
 
     // Rï¿½cupï¿½ration des donnï¿½es du formulaire
-    const dateJourFerme = this.formCreationJourFerme.get('dateJourFerme').value;
-    const typeJourFerme = this.formCreationJourFerme.get('typeJourFerme').value;
-    const commentaireJourFerme = this.formCreationJourFerme.get('commentaireJourFerme').value;
+    const dateJourFerme = this.formModificationJourFerme.get('dateJourFerme').value;
+    const typeJourFerme = this.formModificationJourFerme.get('typeJourFerme').value;
+    const commentaireJourFerme = this.formModificationJourFerme.get('commentaireJourFerme').value;
 
     // on formate la date du jour au format 'yyyy-MM-dd'
     const dateAujourdhui = formatDate(Date.now(), 'yyyy-MM-dd', 'en-US');
@@ -65,7 +83,7 @@ export class CreationJourFerieComponent implements OnInit {
       this.messageErreur = 'ERREUR. LE COMMENTAIRE EST OBLIGATOIRE POUR LES JOURS FERIES.';
     }
     else {
-      this.jourFermeService.ajouterJourFerme(dateJourFerme, typeJourFerme, commentaireJourFerme).subscribe(
+      this.jourFermeService.modifierJourFerme(this.id, dateJourFerme, typeJourFerme, commentaireJourFerme).subscribe(
         () => { },
         () => {
           this.messageErreur = 'ERREUR';
@@ -74,10 +92,11 @@ export class CreationJourFerieComponent implements OnInit {
           this.messageErreur = '';
           setTimeout(() => {
             // Redirection au bout de 2 secondes
-            this.router.navigate(['listerJourFerie']);
+            this.router.navigate(['listerJourFerme']);
           }, 2000);
         }
       );
     }
   }
+
 }
